Moved ToastContainer to higher App component level
