+++ client/src/components/pages/signin/Signin.tsx
@@ -1,0 +1,26 @@
+import { Box, Typography } from "@mui/material";
+import Navigation from "../../navigation/authNavigation";
+import SigninForm from "components/auth/signin/SigninForm";
+import { Navigate } from "react-router-dom";
+import { isAuth } from "components/auth/helpers/helpers";
+
+type SigninProps = {
+  displayToast: (message: string, type: any) => void;
+};
+
+function Signin({ displayToast }: SigninProps) {
+  const title: string = "SIGNIN";
+
+  return (
+    <Box>
+      {isAuth() ? <Navigate to="/home" /> : null}
+      <Typography sx={{ padding: "20px" }} variant="h4">
+        {title}
+      </Typography>
+      <Navigation />
+      <SigninForm displayToast={displayToast} />
+    </Box>
+  );
+}
+
+export default Signin;
